fix(dropdown): extract DropdownWithContext
